Update API endpoints and permissions
This update involves altering several API endpoints, mainly thread, emoji, and sticker functions. It also introduces new fields for a few types and revises the permissions for various functionalities. Changes are made to account for GuildMedia channels in addition to GuildForum channels. The Emoji and Sticker functions' permissions have also been specified based on who created them. Lastly, the Go version was updated in the workflow and dependencies.